fix(assemble): Add missing template option bindings to jlink
Fix missing template variable binding in jlink assembler.

Add distributionJavaMainModule, distributionJavaMainClass
and distributionJavaMainJar to template context in JlinkAssemblerProcessor
similarly it's done in JavaArchiveAssemblerProcessor.

This fix don't try to add distributionJavaOptions as that's not
used in a jlink template and assembler structure seems to be slightly
different compared to java-archive.

Fixes #1496